Lee y descarga datos meteorologicos de una estacion

Descarga (si es necesario) y lee un archivo CSV con datos meteorologicos correspondientes a una estacion especifica.

Usage

leer_estacion(id_estacion, ruta)

Arguments

id_estacion

Cadena de texto con el identificador de la estacion.

ruta

Cadena de texto indicando donde guardar o leer el archivo CSV.

Value

Un tibble con los datos meteorologicos de la estacion.

Examples

data("NH0472")
head(NH0472)
#> # A tibble: 6 × 35
#>   id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>   <chr>  <date>                        <dbl>                           <dbl>
#> 1 NH0472 1961-07-01                     11.1                            13.4
#> 2 NH0472 1961-07-02                      5.5                            10.5
#> 3 NH0472 1961-07-03                      6                              10.8
#> 4 NH0472 1961-07-04                     NA                              NA  
#> 5 NH0472 1961-07-05                     12.4                            14.8
#> 6 NH0472 1961-07-06                     11                              16.5
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <dbl>,
#> #   temperatura_suelo_5cm_media <lgl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>,
#> #   nieve <dbl>, heliofania_efectiva <dbl>, heliofania_relativa <dbl>, …